+++ packages/base/src/features/story/components/ListStoryTitleBar.tsx
@@ -1,63 +1,16 @@
 import { IJupyterGISModel } from '@jupytergis/schema';
-import React, {
-  useCallback,
-  useLayoutEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import React, { useMemo } from 'react';
 
+import { ListStoryTitleBarDesktop } from '@/src/features/story/components/ListStoryTitleBarDesktop';
+import { ListStoryTitleBarMobile } from '@/src/features/story/components/ListStoryTitleBarMobile';
 import { useListStoryScrollTrackContext } from '@/src/features/story/context/ListStoryScrollTrackContext';
 import { useCurrentSegmentIndex } from '@/src/features/story/hooks/useCurrentSegmentIndex';
-import type { IStorySegmentViewItem } from '@/src/features/story/types/types';
 import { buildStorySegmentViewItems } from '@/src/features/story/utils/storySegmentViewItems';
-import { Button } from '@/src/shared/components/Button';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@/src/shared/components/Popover';
-import { ChevronLeft, ChevronRight, Menu } from 'lucide-react';
 
 interface IListStoryTitleBarProps {
   model: IJupyterGISModel;
   isMobile: boolean;
 }
-
-interface IListStoryTitleBarSegmentsProps {
-  segmentItems: IStorySegmentViewItem[];
-  currentIndex: number;
-  onSegmentClick: (index: number) => void;
-}
-
-const ListStoryTitleBarSegments = React.forwardRef<
-  HTMLDivElement,
-  IListStoryTitleBarSegmentsProps
->(function ListStoryTitleBarSegments(
-  { segmentItems, currentIndex, onSegmentClick },
-  ref,
-) {
-  return (
-    <div ref={ref} className="jgis-story-title-bar-segments">
-      {segmentItems.map(item => {
-        const isActive = item.index === currentIndex;
-        return (
-          <button
-            key={item.id}
-            type="button"
-            className="jGIS-layer-browser-category jgis-story-title-bar-segment"
-            data-state={isActive ? 'active' : 'inactive'}
-            aria-current={isActive ? 'true' : undefined}
-            aria-label={`Go to ${item.layerName}`}
-            onClick={() => onSegmentClick(item.index)}
-          >
-            {item.layerName}
-          </button>
-        );
-      })}
-    </div>
-  );
-});
 
 /**
  * Story stage title bar: one button per segment, labeled with the segment layer name.
@@ -66,9 +19,6 @@ export function ListStoryTitleBar({
   model,
   isMobile,
 }: IListStoryTitleBarProps): JSX.Element {
-  const segmentsRef = useRef<HTMLDivElement>(null);
-  const [hasOverflow, setHasOverflow] = useState(false);
-  const [menuOpen, setMenuOpen] = useState(false);
   const currentIndex = useCurrentSegmentIndex(model);
   const { scrollToSegmentIndex } = useListStoryScrollTrackContext();
 
@@ -78,188 +28,15 @@ export function ListStoryTitleBar({
     [model],
   );
 
-  const activeSegment = segmentItems.find(item => item.index === currentIndex);
+  const titleBarProps = {
+    segmentItems,
+    currentIndex,
+    onSegmentClick: scrollToSegmentIndex,
+  };
 
-  const currentPosition = segmentItems.findIndex(
-    item => item.index === currentIndex,
-  );
-  const prevSegmentIdRef = useRef<string | undefined>(undefined);
-  const slideDirectionRef = useRef<'next' | 'prev' | undefined>(undefined);
-
-  if (activeSegment?.id !== prevSegmentIdRef.current) {
-    const prevId = prevSegmentIdRef.current;
-    if (prevId) {
-      const prevPosition = segmentItems.findIndex(item => item.id === prevId);
-      if (prevPosition >= 0 && currentPosition >= 0) {
-        slideDirectionRef.current =
-          currentPosition > prevPosition ? 'next' : 'prev';
-      } else {
-        slideDirectionRef.current = undefined;
-      }
-    } else {
-      slideDirectionRef.current = undefined;
-    }
-    prevSegmentIdRef.current = activeSegment?.id;
+  if (isMobile) {
+    return <ListStoryTitleBarMobile {...titleBarProps} />;
   }
 
-  const hasPrev = currentPosition > 0;
-  const hasNext =
-    currentPosition >= 0 && currentPosition < segmentItems.length - 1;
-
-  const goToAdjacentSegment = useCallback(
-    (direction: -1 | 1): void => {
-      const nextPosition = currentPosition + direction;
-      const nextItem = segmentItems[nextPosition];
-      if (!nextItem) {
-        return;
-      }
-
-      scrollToSegmentIndex(nextItem.index);
-    },
-    [currentPosition, segmentItems, scrollToSegmentIndex],
-  );
-
-  const handleMenuSegmentClick = useCallback(
-    (index: number): void => {
-      scrollToSegmentIndex(index);
-      setMenuOpen(false);
-    },
-    [scrollToSegmentIndex],
-  );
-
-  useLayoutEffect(() => {
-    if (isMobile) {
-      return;
-    }
-
-    const segments = segmentsRef.current;
-    if (!segments) {
-      return;
-    }
-
-    const update = (): void => {
-      setHasOverflow(segments.scrollWidth > segments.clientWidth);
-    };
-
-    update();
-
-    const ro = new ResizeObserver(update);
-    ro.observe(segments);
-
-    return () => {
-      ro.disconnect();
-    };
-  }, [isMobile]);
-
-  useLayoutEffect(() => {
-    if (isMobile) {
-      return;
-    }
-
-    const segments = segmentsRef.current;
-    if (!segments) {
-      return;
-    }
-
-    const active = segments.querySelector(
-      '.jgis-story-title-bar-segment[data-state="active"]',
-    );
-    if (!(active instanceof HTMLElement)) {
-      return;
-    }
-
-    active.scrollIntoView({
-      behavior: 'smooth',
-      block: 'nearest',
-      inline: 'nearest',
-    });
-  }, [currentIndex, isMobile]);
-
-  return (
-    <nav
-      className={`jgis-story-title-bar${
-        isMobile ? ' jgis-story-title-bar--mobile' : ''
-      }`}
-      aria-label="Story segments"
-    >
-      {isMobile ? (
-        <Popover open={menuOpen} onOpenChange={setMenuOpen}>
-          <PopoverTrigger asChild>
-            <Button
-              type="button"
-              variant="ghost"
-              className="jgis-story-title-bar-menu-btn"
-              aria-label="Open story menu"
-            >
-              <Menu />
-            </Button>
-          </PopoverTrigger>
-          <PopoverContent
-            align="center"
-            side="bottom"
-            className="jgis-story-title-bar-segment-menu"
-          >
-            {segmentItems.map(item => {
-              const isActive = item.index === currentIndex;
-              return (
-                <button
-                  key={item.id}
-                  type="button"
-                  className="jGIS-layer-browser-category jgis-story-title-bar-segment-menu-item"
-                  data-state={isActive ? 'active' : 'inactive'}
-                  aria-current={isActive ? 'true' : undefined}
-                  onClick={() => handleMenuSegmentClick(item.index)}
-                >
-                  {item.layerName}
-                </button>
-              );
-            })}
-          </PopoverContent>
-        </Popover>
-      ) : null}
-      {!isMobile && hasOverflow ? (
-        <Button
-          type="button"
-          variant="ghost"
-          size="icon-sm"
-          className="jgis-story-title-bar-scroll-btn"
-          aria-label="Previous segment"
-          disabled={!hasPrev}
-          onClick={() => goToAdjacentSegment(-1)}
-        >
-          <ChevronLeft />
-        </Button>
-      ) : null}
-      {isMobile ? (
-        <span
-          key={activeSegment?.id}
-          className="jGIS-layer-browser-category jgis-story-title-bar-active-segment"
-          data-state="active"
-          data-slide-direction={slideDirectionRef.current}
-        >
-          {activeSegment?.layerName ?? ''}
-        </span>
-      ) : (
-        <ListStoryTitleBarSegments
-          ref={segmentsRef}
-          segmentItems={segmentItems}
-          currentIndex={currentIndex}
-          onSegmentClick={scrollToSegmentIndex}
-        />
-      )}
-      {!isMobile && hasOverflow ? (
-        <Button
-          type="button"
-          variant="ghost"
-          size="icon-sm"
-          className="jgis-story-title-bar-scroll-btn"
-          aria-label="Next segment"
-          disabled={!hasNext}
-          onClick={() => goToAdjacentSegment(1)}
-        >
-          <ChevronRight />
-        </Button>
-      ) : null}
-    </nav>
-  );
+  return <ListStoryTitleBarDesktop {...titleBarProps} />;
 }
